+++ crates/ryo3-serde/src/ser/py_types/py_enum.rs
@@ -1,0 +1,36 @@
+use pyo3::{Borrowed, PyAny, intern, prelude::*};
+use serde::ser::{Serialize, Serializer};
+
+use crate::{Depth, PyAnySerializer, errors::pyerr2sererr, ser::PySerializeContext};
+
+pub(crate) struct PyEnumSerializer<'a, 'py> {
+    ctx: PySerializeContext<'py>,
+    obj: Borrowed<'a, 'py, PyAny>,
+    depth: Depth,
+}
+
+impl<'a, 'py> PyEnumSerializer<'a, 'py> {
+    #[inline]
+    pub(crate) fn new(
+        obj: Borrowed<'a, 'py, PyAny>,
+        ctx: PySerializeContext<'py>,
+        depth: Depth,
+    ) -> Self {
+        Self { ctx, obj, depth }
+    }
+}
+
+impl Serialize for PyEnumSerializer<'_, '_> {
+    #[inline]
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let value = self
+            .obj
+            .getattr(intern!(self.obj.py(), "value"))
+            .map_err(pyerr2sererr)?;
+        PyAnySerializer::new_with_depth(value.as_borrowed(), self.ctx, self.depth)
+            .serialize(serializer)
+    }
+}